Move wrapping get and reuse in glyf.rs
